HAJ-1 Update Code Tawakkalna Job After Testing

+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/dto/TawakkalnaApplicantInputDto.java
@@ -18,7 +18,7 @@ public class TawakkalnaApplicantInputDto implements Serializable {
     private String passportNumber;
 
     @JsonProperty("nationalitycode")
-    private int nationalityCode;
+    private long nationalityCode;
 
     @JsonProperty("smartcardnumber")
     private String smartCardNumber;
